Add sealed class "Result".
