--- app/src/main/java/com/example/dreamwallpaper/domain/ResourcesProvider.kt
@@ -1,12 +1,0 @@
-package com.example.dreamwallpaper.domain
-
-import android.content.Context
-import androidx.annotation.StringRes
-
-class ResourcesProvider(
-    private val context: Context
-    ) {
-    fun getString(@StringRes stringResId: Int): String {
-        return context.getString(stringResId)
-    }
-}
