+++ app/src/main/java/com/example/dreamwallpaper/screens/main/MainActivity.kt
@@ -12,6 +12,7 @@ import androidx.navigation.ui.AppBarConfiguration
 import androidx.navigation.ui.setupWithNavController
 import com.example.dreamwallpaper.R
 import com.example.dreamwallpaper.databinding.ActivityMainBinding
+import com.example.dreamwallpaper.util.setTitle
 import com.google.android.material.appbar.MaterialToolbar
 import java.util.regex.Pattern
 
@@ -45,27 +46,6 @@ class MainActivity : AppCompatActivity() {
             }
         }
 
-    }
-
-    fun MaterialToolbar.setTitle(label: CharSequence?, textView: TextView, arguments: Bundle?) {
-        if (label != null) {
-            // Fill in the data pattern with the args to build a valid URI
-            val title = StringBuffer()
-            val fillInPattern = Pattern.compile("\\{(.+?)\\}")
-            val matcher = fillInPattern.matcher(label)
-            while (matcher.find()) {
-                val argName = matcher.group(1)
-                if (arguments != null && arguments.containsKey(argName)) {
-                    matcher.appendReplacement(title, "")
-                    title.append(arguments.get(argName).toString())
-                } else {
-                    return //returning because the argument required is not found
-                }
-            }
-            matcher.appendTail(title)
-            setTitle("")
-            textView.text = title
-        }
     }
 
     override fun onBackPressed() {
