+++ app/src/main/res/layout/activity_main.xml
@@ -15,9 +15,10 @@
         <com.google.android.material.appbar.CollapsingToolbarLayout
             android:id="@+id/collapsing_toolbar_layout"
             android:layout_width="match_parent"
-            android:layout_height="100dp"
+            android:layout_height="70dp"
             android:fitsSystemWindows="true"
-            app:layout_scrollFlags="scroll|enterAlways|enterAlwaysCollapsed">
+            android:background="?attr/colorPrimary"
+            app:layout_scrollFlags="scroll|exitUntilCollapsed|snap">
 
             <com.google.android.material.appbar.MaterialToolbar
                 android:id="@+id/toolbar"
@@ -33,6 +34,7 @@
                     android:layout_height="wrap_content"
                     android:text="@string/app_name"
                     android:textSize="24sp"
+                    app:layout_collapseMode="parallax"
                     android:textStyle="bold"
                     android:textColor="?attr/colorPrimaryVariant"
                     android:fontFamily="@font/tt_norms_regular"/>
