+++ app/src/main/res/layout/fragment_image_list.xml
@@ -22,7 +22,8 @@
         android:contentDescription="Tab"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        app:background="@color/black"
+        android:backgroundTint="?attr/colorPrimary"
+        app:layout_collapseMode="pin"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
